@fpetkovski 2022-10-07
Yes, but `sample` was returning a new function every time it was called 🤦